Address comments - refactoring
