+++ opensearch-observability/src/main/kotlin/org/opensearch/observability/collaboration/action/CollaborationActions.kt
@@ -3,13 +3,13 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package org.opensearch.observability.action
+package org.opensearch.observability.collaboration.action
 
 import org.opensearch.OpenSearchStatusException
 import org.opensearch.commons.authuser.User
 import org.opensearch.observability.ObservabilityPlugin
+import org.opensearch.observability.collaboration.model.CollaborationObjectDoc
 import org.opensearch.observability.index.CollaborationIndex
-import org.opensearch.observability.model.CollaborationObjectDoc
 import org.opensearch.observability.security.UserAccessManager
 import org.opensearch.observability.util.logger
 import org.opensearch.rest.RestStatus
