+++ opensearch-observability/src/main/kotlin/org/opensearch/observability/collaboration/action/CreateCollaborationObjectAction.kt
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package org.opensearch.observability.action
+package org.opensearch.observability.collaboration.action
 
 import org.opensearch.action.ActionType
 import org.opensearch.action.support.ActionFilters
@@ -11,6 +11,7 @@ import org.opensearch.client.Client
 import org.opensearch.common.inject.Inject
 import org.opensearch.common.xcontent.NamedXContentRegistry
 import org.opensearch.commons.authuser.User
+import org.opensearch.observability.action.PluginBaseAction
 import org.opensearch.transport.TransportService
 
 /**
