+++ opensearch-observability/src/main/kotlin/org/opensearch/observability/collaboration/action/CreateCollaborationObjectResponse.kt
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package org.opensearch.observability.action
+package org.opensearch.observability.collaboration.action
 
 import org.opensearch.common.io.stream.StreamInput
 import org.opensearch.common.io.stream.StreamOutput
