+++ opensearch-observability/src/main/kotlin/org/opensearch/observability/collaboration/model/CollaborationTypeDataProperties.kt
@@ -3,10 +3,12 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package org.opensearch.observability.model
+package org.opensearch.observability.collaboration.model
 
 import org.opensearch.common.io.stream.Writeable
 import org.opensearch.common.xcontent.XContentParser
+import org.opensearch.observability.model.BaseObjectData
+import org.opensearch.observability.model.XParser
 
 internal object CollaborationTypeDataProperties {
     /**
